LayoutTests/imported/w3c/web-platform-tests/fetch/api/cors/cors-preflight.any.worker-expected.txt

CONSOLE MESSAGE: Method DELETE is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Method PUT is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Method PATCH is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Method NEW is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Request header field x-test-header1 is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: Request header field Content-Type is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: Method PUT is not allowed by Access-Control-Allow-Methods.

PASS CORS [DELETE], server allows 
PASS CORS [DELETE], server refuses 
PASS CORS [PUT], server allows 
PASS CORS [PUT], server refuses 
PASS CORS [PATCH], server allows 
PASS CORS [PATCH], server refuses 
PASS CORS [NEW], server allows 
PASS CORS [NEW], server refuses 
PASS CORS [GET] [x-test-header: allowed], server allows 
PASS CORS [GET] [x-test-header: refused], server refuses 
PASS CORS [GET] [several headers], server allows 
PASS CORS [GET] [several headers], server refuses 
PASS CORS [PUT] [several headers], server allows 
PASS CORS [PUT] [several headers], server refuses 
PASS CORS [PUT] [only safe headers], server allows 
